+++ SciProfile/PSModuleUtils/Functions/Set-Window.ps1
@@ -78,14 +78,14 @@ Function Set-Window {
     Description: Display the coordinates of the window for the current 
                  PowerShell session in a table view.
     #>
-    [CmdletBinding(PositionalBinding=$True, DefaultParameterSetName="Name")]
+    [CmdletBinding(PositionalBinding, DefaultParameterSetName="Name")]
 
     Param (
 
         [Parameter(Mandatory=$False, ValueFromPipelineByPropertyName=$True, ParameterSetName="Name", HelpMessage="Name of the process to determine the window characteristics. (All processes if omitted).")]
         [String]$ProcessName="*",
 
-        [Parameter(Mandatory=$True, ValueFromPipeline=$False,       ParameterSetName="Id", HelpMessage="Id of the process to determine the window characteristics. ")]
+        [Parameter(Mandatory, ValueFromPipeline=$False,       ParameterSetName="Id", HelpMessage="Id of the process to determine the window characteristics. ")]
         [Int]$Id,
 
         [Parameter(HelpMessage="Set the position of the window in pixels from the left.")]