+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.3.1](https://github.com/wbrandenburger/SCiProfile/tree/0.3.1) (2019-09-15)
+
+**Implemented enhancements:**
+
+- Added function `Set-ProfileProjectList`
+
 ## [0.3.0](https://github.com/wbrandenburger/SCiProfile/tree/0.3.0) (2019-09-12)
 
 **Implemented enhancements:**